Moved congestion charge details to separate file that the pdf and parcels folders both access
